updated parse-need to use the content extraction parsers from the defined detail-definitions of all present usecases
