+++ webofneeds/won-owner-webapp/src/main/webapp/config/detail-definitions.js
@@ -2,7 +2,8 @@
 // TODO: each detail picker should know it's own rdf template
 // --> both for parsing to and from rdf
 // --> templates are used in need-builder (toRDF) and in parse-need (from RDF)
-import { getIn } from "../app/utils.js";
+import { getIn, is } from "../app/utils.js";
+import Immutable from "immutable";
 
 export const details = {
   description: {
@@ -15,6 +16,9 @@ export const details = {
         return { "dc:description": undefined };
       }
       return { "dc:description": value };
+    },
+    parseFromRDF: function(jsonLDImm) {
+      return jsonLDImm && jsonLDImm.get("dc:description");
     },
   },
   location: {
@@ -57,6 +61,112 @@ export const details = {
         },
       };
     },
+    parseFromRDF: function(jsonLDImm) {
+      const jsonldLocation = jsonLDImm && jsonLDImm.get("won:hasLocation");
+      if (!jsonldLocation) return undefined; // NO LOCATION PRESENT
+
+      const jsonldLocationImm = Immutable.fromJS(jsonldLocation);
+
+      let location = {
+        address: undefined,
+        lat: undefined,
+        lng: undefined,
+        nwCorner: {
+          lat: undefined,
+          lng: undefined,
+        },
+        seCorner: {
+          lat: undefined,
+          lng: undefined,
+        },
+      };
+
+      location.address =
+        jsonldLocationImm.get("s:name") ||
+        jsonldLocationImm.get("http://schema.org/name");
+
+      location.lat = Number.parseFloat(
+        jsonldLocationImm.getIn(["s:geo", "s:latitude"]) ||
+          jsonldLocationImm.getIn([
+            "http://schema.org/geo",
+            "http://schema.org/latitude",
+          ])
+      );
+      location.lng = Number.parseFloat(
+        jsonldLocationImm.getIn(["s:geo", "s:longitude"]) ||
+          jsonldLocationImm.getIn([
+            "http://schema.org/geo",
+            "http://schema.org/longitude",
+          ])
+      );
+
+      location.nwCorner.lat = Number.parseFloat(
+        jsonldLocationImm.getIn([
+          "won:hasBoundingBox",
+          "won:hasNorthWestCorner",
+          "s:latitude",
+        ]) ||
+          jsonldLocationImm.getIn([
+            "won:hasBoundingBox",
+            "won:hasNorthWestCorner",
+            "http://schema.org/latitude",
+          ])
+      );
+      location.nwCorner.lng = Number.parseFloat(
+        jsonldLocationImm.getIn([
+          "won:hasBoundingBox",
+          "won:hasNorthWestCorner",
+          "s:longitude",
+        ]) ||
+          jsonldLocationImm.getIn([
+            "won:hasBoundingBox",
+            "won:hasNorthWestCorner",
+            "http://schema.org/longitude",
+          ])
+      );
+      location.seCorner.lat = Number.parseFloat(
+        jsonldLocationImm.getIn([
+          "won:hasBoundingBox",
+          "won:hasSouthEastCorner",
+          "s:latitude",
+        ]) ||
+          jsonldLocationImm.getIn([
+            "won:hasBoundingBox",
+            "won:hasSouthEastCorner",
+            "http://schema.org/latitude",
+          ])
+      );
+      location.seCorner.lng = Number.parseFloat(
+        jsonldLocationImm.getIn([
+          "won:hasBoundingBox",
+          "won:hasSouthEastCorner",
+          "s:longitude",
+        ]) ||
+          jsonldLocationImm.getIn([
+            "won:hasBoundingBox",
+            "won:hasSouthEastCorner",
+            "http://schema.org/longitude",
+          ])
+      );
+
+      if (
+        location.address &&
+        location.lat &&
+        location.lng &&
+        location.nwCorner.lat &&
+        location.nwCorner.lng &&
+        location.seCorner.lat &&
+        location.seCorner.lng
+      ) {
+        return Immutable.fromJS(location);
+      }
+
+      console.error(
+        "Cant parse location, data is an invalid location-object: ",
+        jsonldLocationImm.toJS()
+      );
+      return undefined;
+    },
   },
   person: {
     identifier: "person",
@@ -87,6 +197,46 @@ export const details = {
           ? getIn(value, ["skills"]).toJS()
           : undefined,
       };
+    },
+    parseFromRDF: function(jsonLDImm) {
+      if (!jsonLDImm) return undefined;
+
+      let person = {
+        name: undefined,
+        title: undefined,
+        company: undefined,
+        position: undefined,
+        skills: undefined,
+        // bio: undefined,
+      };
+
+      person.name = jsonLDImm.get("foaf:name");
+      person.title = jsonLDImm.get("foaf:title");
+      person.company = jsonLDImm.get("s:worksFor");
+      person.position = jsonLDImm.get("s:jobTitle");
+      person.skills = jsonLDImm.get("s:knowsAbout")
+        ? Immutable.List.isList(jsonLDImm.get("s:knowsAbout"))
+          ? jsonLDImm.get("s:knowsAbout")
+          : Immutable.List.of(jsonLDImm.get("s:knowsAbout"))
+        : undefined;
+      //person.bio = isOrSeeksImm.get("dc:description");
+
+      // if there's anything, use it
+      if (
+        person.name ||
+        person.title ||
+        person.company ||
+        person.position ||
+        person.skills
+      ) {
+        return Immutable.fromJS(person);
+      }
+
+      console.error(
+        "Cant parse person, data does not contain enough information: ",
+        jsonLDImm.toJS()
+      );
+      return undefined;
     },
   },
   // TODO: rename to travelAction
@@ -127,6 +277,88 @@ export const details = {
         },
       };
     },
+    parseFromRDF: function(jsonLDImm) {
+      const jsonTravelAction = jsonLDImm && jsonLDImm.get("won:travelAction");
+      if (!jsonTravelAction) return undefined;
+
+      const travelActionImm = Immutable.fromJS(jsonTravelAction);
+
+      let travelAction = {
+        fromAddress: undefined,
+        fromLocation: {
+          lat: undefined,
+          lng: undefined,
+        },
+        toAddress: undefined,
+        toLocation: {
+          lat: undefined,
+          lng: undefined,
+        },
+      };
+
+      travelAction.fromAddress =
+        travelActionImm.getIn(["s:fromLocation", "s:name"]) ||
+        travelActionImm.getIn([
+          "http://schema.org/fromLocation",
+          "http://schema.org/name",
+        ]);
+
+      travelAction.fromLocation.lat =
+        travelActionImm.getIn(["s:fromLocation", "s:geo", "s:latitude"]) ||
+        travelActionImm.getIn([
+          "http://schema.org/fromLocation",
+          "http://schema.org/geo",
+          "http://schema.org/latitude",
+        ]);
+
+      travelAction.fromLocation.lng =
+        travelActionImm.getIn(["s:fromLocation", "s:geo", "s:longitude"]) ||
+        travelActionImm.getIn([
+          "http://schema.org/fromLocation",
+          "http://schema.org/geo",
+          "http://schema.org/longitude",
+        ]);
+
+      travelAction.toAddress =
+        travelActionImm.getIn(["s:toLocation", "s:name"]) ||
+        travelActionImm.getIn([
+          "http://schema.org/toLocation",
+          "http://schema.org/name",
+        ]);
+
+      travelAction.toLocation.lat =
+        travelActionImm.getIn(["s:toLocation", "s:geo", "s:latitude"]) ||
+        travelActionImm.getIn([
+          "http://schema.org/toLocation",
+          "http://schema.org/geo",
+          "http://schema.org/latitude",
+        ]);
+
+      travelAction.toLocation.lng =
+        travelActionImm.getIn(["s:toLocation", "s:geo", "s:longitude"]) ||
+        travelActionImm.getIn([
+          "http://schema.org/toLocation",
+          "http://schema.org/geo",
+          "http://schema.org/longitude",
+        ]);
+
+      if (
+        (travelAction.fromAddress &&
+          travelAction.fromLocation.lat &&
+          travelAction.fromLocation.lng) ||
+        (travelAction.toAddress &&
+          travelAction.toLocation.lat &&
+          travelAction.toLocation.lng)
+      ) {
+        return Immutable.fromJS(travelAction);
+      }
+
+      console.error(
+        "Cant parse travelAction, data is an invalid travelAction-object: ",
+        travelActionImm.toJS()
+      );
+      return undefined;
+    },
   },
   tags: {
     identifier: "tags",
@@ -139,6 +371,11 @@ export const details = {
       }
       return { "won:hasTag": value };
     },
+    parseFromRDF: function({ jsonLD }) {
+      //TODO: Implement this method
+      console.log("trying to parse RDF to content from: ", jsonLD);
+      return undefined;
+    },
   },
   ttl: {
     identifier: "ttl",
@@ -150,6 +387,26 @@ export const details = {
         return undefined;
       }
       // TODO: return value
+    },
+    parseFromRDF: function(jsonLDImm) {
+      const tags = jsonLDImm && jsonLDImm.get("won:hasTags");
+
+      if (!tags) {
+        return undefined;
+      } else if (is("String", tags)) {
+        return Immutable.fromJS([tags]);
+      } else if (is("Array", tags)) {
+        return Immutable.fromJS(tags);
+      } else if (Immutable.List.isList(tags)) {
+        return tags; // id; it is already in the format we want
+      } else {
+        console.error(
+          "Found unexpected format of tags (should be Array, " +
+            "Immutable.List, or a single tag as string): " +
+            JSON.stringify(tags)
+        );
+        return undefined;
+      }
     },
   },
 };
